# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTrack",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of AquaTrack app"
  },
  "servers": [
    {
      "url": "https://aquatrack-01.onrender.com"
    }
  ],
...
  "paths": {
    "/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "registerUser",
        "summary": "User Registration",
        "description": "Registers a new user by providing their email and password. On successful registration, a 201 Created status is returned along with the user data. If the email is already in use, a 409 Conflict error is returned.",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "description": "User's email",
                    "type": "string",
                    "example": "john.doe@example.com"
                  },
                  "password": {
                    "description": "User's password",
                    "type": "string",
                    "example": "12345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully registered a user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email address is already in use",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "ConflictError"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Email in use"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "loginUser",
        "summary": "User Authentication",
        "description": "Authenticates a user by validating their email and password. On successful authentication, a session cookie and a refresh token cookie are set, and an access token is returned. If the credentials are incorrect, a 401 Unauthorized error is returned.",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "headers": {
              "Set-Cookie": {
                "description": "Session cookie",
                "schema": {
                  "type": "string",
                  "example": "sessionId=j%3A%226777d431ab6c2bc095d944a5%22; Path=/; Expires=Sun, 02 Feb 2025 12:12:33 GMT; HttpOnly"
                }
              },
              "\u0000Set-Cookie": {
                "description": "Refresh token cookie",
                "schema": {
                  "type": "string",
                  "example": "refreshToken=ijKGicDPD8fz2egibIhkNGQkj2ziaQXIURK8jEyA; Path=/; Expires=Sun, 02 Feb 2025 12:12:33 GMT; HttpOnly"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in a user!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "accessToken",
                        "user"
                      ],
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "Qu469Tkd8Y2mAPIBo96le7ODiupGgjYI+2D2FBOO"
                        },
                        "user": {
                          "$ref": "#/components/schemas/user"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "UnauthorizedError"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Email or password is incorrect"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "logoutUser",
        "summary": "User Logout",
        "description": "Logs out the authenticated user by invalidating their bearer token. A 204 No Content response is returned upon successful logout. If the token is invalid or expired, a 401 Unauthorized error is returned.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully logged out the user"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/users/update": {
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "updateUser",
        "summary": "Update User Account Information",
        "description": "Updates a user's account details, including name, gender, weight, active time in sports, and daily water intake. Returns a confirmation message and updated user data upon success, or a 401 Unauthorized error if access is denied.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "User's name",
                    "type": "string",
                    "example": "John Doe"
                  },
                  "avatar": {
                    "description": "User's avatar",
                    "type": "string",
                    "format": "binary"
                  },
                  "gender": {
                    "description": "User's gender",
                    "type": "string",
                    "enum": [
                      "woman",
                      "man"
                    ],
                    "example": "man"
                  },
                  "weight": {
                    "description": "User's weight (in kilograms)",
                    "type": "number",
                    "example": 80
                  },
                  "activeTime": {
                    "description": "User's time of active participation in sports (in hours)",
                    "type": "number",
                    "example": 1
                  },
                  "dailyNorm": {
                    "description": "User's daily norm of water intake (in milliliters)",
                    "type": "number",
                    "example": 3000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated a user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "description": "User's id (automatically generated)",
                          "type": "string",
                          "example": "677eb7405c84945af390b81f"
                        },
                        "email": {
                          "description": "User's email",
                          "type": "string",
                          "example": "john.doe@example.com"
                        },
                        "name": {
                          "description": "Users' name",
                          "type": "string",
                          "example": "john Doe"
                        },
                        "gender": {
                          "description": "User's gender",
                          "type": "string",
                          "enum": [
                            "woman",
                            "man"
                          ],
                          "example": "man"
                        },
                        "weight": {
                          "description": "User's weight (in kilograms)",
                          "type": "number",
                          "example": 80
                        },
                        "activeTime": {
                          "description": "User's time of active participation in sports (in hours)",
                          "type": "number",
                          "example": 1
                        },
                        "dailyNorm": {
                          "description": "User's daily norm of water intake (in milliliters)",
                          "type": "number",
                          "example": 3000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "_id": {
            "description": "User's id (automatically generated)",
            "type": "string",
            "example": "677eb7405c84945af390b81f"
          },
          "email": {
            "description": "User's email",
            "type": "string",
            "example": "john.doe@example.com"
          },
          "name": {
            "description": "Users' name",
            "type": "string",
            "example": "john.doe",
            "default": "Local part of email address"
          },
          "avatar": {
            "description": "Link to user's avatar",
            "type": "string",
            "example": "https://res.cloudinary.com/dxwxmlz4k/image/upload/v1736940725/oahu8xtkfimt793xz1ce.jpg"
          },
          "gender": {
            "description": "User's gender",
            "type": "string",
            "enum": [
              "woman",
              "man"
            ],
            "example": "woman",
            "default": "woman"
          },
          "weight": {
            "description": "User's weight (in kilograms)",
            "type": "number",
            "example": 0,
            "default": 0
          },
          "activeTime": {
            "description": "User's time of active participation in sports (in hours)",
            "type": "number",
            "example": 0,
            "default": 0
          },
          "dailyNorm": {
            "description": "User's daily norm of water intake (in milliliters)",
            "type": "number",
            "example": 1500,
            "default": 1500
          }
        }
      },
...
    },
    "responses": {
...
      "401": {
        "description": "Invalid or expired access token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}